updated configrations

+++ train.py
@@ -75,7 +75,7 @@ def main():
         load_in_4bit=True,
         bnb_4bit_use_double_quant=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.bfloat16,
+        bnb_4bit_compute_dtype=torch.float16,
     )
 
     model=AutoModelForCausalLM.from_pretrained(
@@ -120,7 +120,7 @@ def main():
             save_steps=500,
             save_total_limit=3,
 
-            bf16=True,
+            fp16=True,
         )
     )
